handled cors, handled negative integer and float, handled query param

+++ taskone/models.py
@@ -4,7 +4,7 @@ from django.db import models
 
 # Create your models here.
 class NumberProperty(models.Model):
-    number = models.BigIntegerField(primary_key=True)
+    number = models.IntegerField(primary_key=True)
     is_prime = models.BooleanField(default=False)
     is_perfect = models.BooleanField(default=False)
